backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
pydantic==2.5.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
pytz==2023.3